extract chat input to new component
