+++ frontend/src/routes/(app)/+page.svelte
@@ -1,9 +1,7 @@
 <script lang="ts">
     import {commentStore} from "$lib/stores/comment.store";
-    import {CommentModel} from "$lib/models/classes/comment.model";
+    import type {CommentModel} from "$lib/models/classes/comment.model";
     import {currentUser, userStore} from "$lib/stores/user.store";
-    import {ChatMessageModel} from "$lib/models/classes/chatMessage.model.js";
-    import {chatMsgCat} from "$lib/models/enums/chatMsgCat";
     import ChatBubble from "$lib/components/chat/ChatBubble.svelte";
     import Modal from "$lib/components/Modal.svelte";
     import {UserModel} from "$lib/models/classes/user.model";
@@ -11,66 +9,16 @@
     import {commentQueue} from "$lib/stores/commentQueue.store";
     import CommentQueue from "$lib/components/chat/CommentQueue.svelte";
     import ConnectionSpinner from "$lib/components/chat/ConnectionSpinner.svelte";
-    import { scale } from 'svelte/transition';
-    import {quintInOut} from 'svelte/easing';
-    import Send from "svelte-material-icons/Send.svelte";
-    import CloseCircleOutline from "svelte-material-icons/CloseCircleOutline.svelte";
+    import ChatInputForm from "$lib/components/forms/ChatInputForm.svelte";
+    import {replyComment} from "$lib/stores/replyComment.store";
 
     let openModal:VoidFunction
     let closeModal:VoidFunction
     let userWithActiveAvatar:UserModel = new UserModel()
-    let replyComment:CommentModel = new CommentModel()
 
     const openAvatarModal = (user:UserModel) => {
         userWithActiveAvatar = user
         openModal()
-    }
-
-    const replyToComment = (comment:CommentModel) => {
-        replyComment = comment
-    }
-
-    const closeReply = () => {
-        replyComment = new CommentModel()
-    }
-
-    function sendMsg() {
-        const input = document.getElementById("msg")! as HTMLInputElement;
-
-        if(input.value === "" || input.value === null) {
-            return
-        }
-
-        const comment = new ChatMessageModel<CommentModel>(
-            chatMsgCat.COMMENT,
-            new CommentModel(
-                input.value,
-                $currentUser.id,
-                replyCommentOrNull(),
-                null,
-                true
-            )
-        )
-
-        commentQueue.addComment(comment)
-        closeReply()
-        input.value = ""
-        input.style.height = "40px"
-        input.focus()
-    }
-
-    function replyCommentOrNull(){
-        return replyComment.createdAt != null ? replyComment : null
-    }
-
-    function sendMsgWithKeyboard(e:KeyboardEvent){
-        const input = e.target as HTMLInputElement
-        input.style.height = "1px"
-        input.style.height = (4+input.scrollHeight)+"px"
-
-        if(e.key == "Enter"){
-            sendMsg()
-        }
     }
 </script>
 
@@ -92,36 +40,11 @@
             <ChatBubble comment={comment}
                         user={$userStore[comment.userId]}
                         isCurrentUser={($currentUser.id === comment.userId)}
-                        openAvatarModal={openAvatarModal}
-                        replyToComment={replyToComment}/>
+                        openAvatarModal={openAvatarModal}/>
         {/each}
     </div>
 
-    <div>
-        {#if replyComment.text !== undefined}
-            <div transition:scale={{ duration: 500, opacity: 0.5, easing: quintInOut }} class="bg-canvas-secondary p-2 mb-1 rounded text-typography-main text-sm relative">
-                <button class="bg-transparent absolute top-1 right-1"  on:click={closeReply}>
-                    <CloseCircleOutline />
-                </button>
-
-                {#if replyComment.userId !== undefined}
-                    <div class="pb-2">
-                        {$userStore[replyComment.userId].name}
-                    </div>
-                {/if}
-                {replyComment.text}
-            </div>
-        {/if}
-        <div class="flex">
-            <textarea class="h-10 text-sm overflow-hidden" name="msg" id="msg" on:keyup={e => sendMsgWithKeyboard(e)}></textarea>
-            <div class="flex flex-col-reverse ml-2">
-                <button on:click={sendMsg}>
-                    <Send size="1.4em"/>
-                </button>
-            </div>
-        </div>
-    </div>
-
+    <ChatInputForm />
 </div>
